+++ src/io.rs
@@ -1,13 +1,13 @@
 use atty::Stream;
 use color_eyre::{eyre::Context, Help, Report};
 use csv::Reader;
-use daggy::NodeIndex;
 use dialoguer::Confirm;
-use log::{debug, warn};
+use std::process;
 use std::{ffi::OsStr, fs::File, fs::OpenOptions, io, path::PathBuf};
+use tracing::instrument;
 
-use self::newick::write_newick;
-use crate::tree::SpideogTree;
+use self::{newick::write_newick, report::ParseKrakenReport};
+use crate::tree::TaxonomyTree;
 
 pub mod newick;
 pub mod report;
@@ -22,16 +22,19 @@ custom_derive! {
     }
 }
 
-pub fn read_report_tree<P>(path: P, headers: bool) -> Result<(SpideogTree, NodeIndex), Report>
+pub fn read_report_tree<P>(path: P, headers: bool) -> Result<TaxonomyTree, Report>
 where
     P: Into<PathBuf>,
 {
-    let path = path.into();
-    let mut reader = get_reader(&path, headers)
-        .wrap_err_with(|| format!("Failed to read file `{}`", path.display()))?;
-    report::read_kraken_report_tree(&mut reader)
-        .wrap_err_with(|| format!("Failed to parse file `{}`", path.display()))
-        .suggestion("Try using the `--has-headers` option if your Kraken report has headers")
+    #[instrument]
+    fn internal_read_report_tree(path: PathBuf, headers: bool) -> Result<TaxonomyTree, Report> {
+        let mut reader = get_reader(&path, headers)
+            .wrap_err_with(|| format!("Failed to read file `{}`", path.display()))?;
+        ParseKrakenReport::parse(&mut reader)
+            .wrap_err_with(|| format!("Failed to parse file `{}`", path.display()))
+            .suggestion("Try using the `--has-headers` option if your Kraken report has headers")
+    }
+    internal_read_report_tree(path.into(), headers)
 }
 
 /* --------------------------------- OUTPUT --------------------------------- */
@@ -44,10 +47,11 @@ custom_derive! {
     }
 }
 
+#[instrument]
 pub fn get_output_file_name(input: &PathBuf, prefix: &Option<String>) -> PathBuf {
     let stem = input
         .file_stem()
-        .unwrap_or_else(|| &OsStr::new("kraken"))
+        .unwrap_or_else(|| OsStr::new("kraken"))
         .to_string_lossy();
     let new_path = if let Some(prefix) = prefix {
         format!("{}_{}.tree", prefix, stem)
@@ -58,8 +62,7 @@ pub fn get_output_file_name(input: &PathBuf, prefix: &Option<String>) -> PathBuf
 }
 
 pub fn write_tree(
-    tree: SpideogTree,
-    root: NodeIndex,
+    tree: TaxonomyTree,
     output: &PathBuf,
     format: &OutputTreeFormat,
     overwrite: bool,
@@ -69,7 +72,7 @@ pub fn write_tree(
 
     match format {
         OutputTreeFormat::Newick => {
-            write_newick(&mut writer, tree, root).wrap_err("Newick Serializer failed")?
+            write_newick(&mut writer, tree).wrap_err("Newick Serializer failed")?
         }
     }
 
@@ -80,31 +83,39 @@ pub fn can_open_file<P>(path: P, overwrite: bool) -> Result<(), Report>
 where
     P: Into<PathBuf>,
 {
-    let path = path.into();
-    if path.exists() {
-        debug!("File `{}` already exists", path.display());
-        if overwrite {
-            warn!("Force overwriting `{}`", path.display());
-        } else if atty::is(Stream::Stdout)
-            && Confirm::new()
-                .with_prompt(format!("Overwrite `{}`?", path.display()))
-                .interact()?
-        {
-            debug!("overwriting file")
+    #[instrument]
+    fn internal_can_open_file(path: PathBuf, overwrite: bool) -> Result<(), Report> {
+        if path.exists() {
+            log::debug!("`{}` already exists", path.display());
+            if overwrite {
+                log::debug!("Force overwriting `{}`", path.display());
+            } else if atty::is(Stream::Stdout) {
+                if Confirm::new()
+                    .with_prompt(format!("Overwrite `{}`?", path.display()))
+                    .interact()?
+                {
+                    log::debug!("overwriting file `{}`", path.display())
+                } else {
+                    log::debug!("refused to overwrite file, exiting...");
+                    process::exit(exitcode::NOPERM);
+                }
+            } else {
+                {
+                    Err(std::io::Error::new(
+                        std::io::ErrorKind::AlreadyExists,
+                        "File already exists",
+                    ))
+                }?
+            }
         } else {
-            {
-                Err(std::io::Error::new(
-                    std::io::ErrorKind::AlreadyExists,
-                    "File already exists",
-                ))
-            }?
+            log::debug!("Will create file `{}`", path.display());
         }
-    } else {
-        debug!("Will create file `{}`", path.display());
+        Ok(())
     }
-    Ok(())
+    internal_can_open_file(path.into(), overwrite)
 }
 
+#[instrument]
 pub fn get_writer(output: &PathBuf, overwrite: bool) -> Result<Box<dyn io::Write>, Report> {
     can_open_file(output, overwrite)
         .wrap_err_with(|| format!("Impossible to write to file `{}`", output.display()))?;
@@ -117,6 +128,7 @@ pub fn get_writer(output: &PathBuf, overwrite: bool) -> Result<Box<dyn io::Write
     ) as Box<dyn io::Write>)
 }
 
+#[instrument]
 pub fn get_reader(input: &PathBuf, headers: bool) -> Result<Reader<File>, csv::Error> {
     csv::ReaderBuilder::new()
         .has_headers(headers)