+++ src/io/newick.rs
@@ -1,28 +1,33 @@
+use color_eyre::Report;
 use daggy::{NodeIndex, Walker};
+use eyre::ContextCompat;
 use std::io;
+use tracing::instrument;
 
-use crate::{tree::SpideogTree, utils::clean_name, ErrorKind};
+use crate::{tree::TaxonomyTree, utils::clean_name};
 
-pub fn write_newick<W>(writer: &mut W, tree: SpideogTree, root: NodeIndex) -> Result<(), ErrorKind>
+pub fn write_newick<W>(writer: &mut W, tree: TaxonomyTree) -> Result<(), Report>
 where
     W: std::io::Write,
 {
-    write_children_recursively(writer, &tree, root, 0)?;
+    write_children_recursively(writer, &tree, tree.origin, 0)?;
     write_end(writer)?;
 
     Ok(())
 }
 
 #[inline]
-pub fn write_name_distance<W>(
-    writer: &mut W,
-    name: String,
-    distance: usize,
-) -> Result<(), io::Error>
+pub fn write_name_distance<W, S>(writer: &mut W, name: S, distance: usize) -> Result<(), io::Error>
 where
     W: io::Write,
+    S: AsRef<str>,
 {
-    write!(writer, "{}:{}", clean_name(name), distance)
+    write!(writer, "{}", format_name_distance(name, distance))
+}
+
+#[inline]
+fn format_name_distance<S: AsRef<str>>(name: S, distance: usize) -> String {
+    format!("{}:{}", clean_name(name.as_ref()), distance)
 }
 
 #[inline]
@@ -30,32 +35,43 @@ pub fn write_end<W>(writer: &mut W) -> Result<(), io::Error>
 where
     W: io::Write,
 {
-    writeln!(writer, ";")
+    write!(writer, "{}", format_end())
+}
+
+#[instrument]
+#[inline]
+fn format_end() -> String {
+    String::from(";\n")
 }
 
 pub fn write_children_recursively<W>(
     writer: &mut W,
-    tree: &SpideogTree,
+    tree: &TaxonomyTree,
     node: NodeIndex,
     parent_indent: usize,
-) -> Result<(), io::Error>
+) -> Result<(), Report>
 where
     W: io::Write,
 {
-    let mut child_walker = tree.children(node);
+    let mut child_walker = tree.tree.children(node);
     let mut children = Vec::new();
-    while let Some((_, node)) = child_walker.walk_next(tree) {
+    while let Some((_, node)) = child_walker.walk_next(&tree.tree) {
         children.push(node);
     }
 
-    let node_data = tree.node_weight(node).unwrap();
+    let node_data = tree.tree.node_weight(node).wrap_err("node not found")?;
+    let distance = node_data
+        .indent
+        .checked_sub(parent_indent)
+        .wrap_err_with(|| {
+            format!(
+                "failed to compute new distance: node {} - parent {}",
+                node_data.indent, parent_indent
+            )
+        })?;
 
     if children.is_empty() {
-        write_name_distance(
-            writer,
-            node_data.organism.name.clone(),
-            node_data.indent - parent_indent,
-        )?;
+        write_name_distance(writer, &node_data.organism.name, distance)?;
     } else {
         writer.write_all(b"(")?;
 
@@ -72,12 +88,23 @@ where
 
         writer.write_all(b")")?;
 
-        write_name_distance(
-            writer,
-            node_data.organism.name.clone(),
-            node_data.indent - parent_indent,
-        )?;
+        write_name_distance(writer, &node_data.organism.name, distance)?;
     }
 
     Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use test_case::test_case;
+
+    #[test_case(&("Homo sapiens", 2), "Homo_sapiens:2")]
+    #[test_case(&("Bacteroidetes/Chlorobi group", 1), "Bacteroidetes_Chlorobi_group:1")]
+    fn test_format_name_distance<S: AsRef<str>>(input: &(S, usize), expected: S) {
+        assert_eq!(
+            format_name_distance(input.0.as_ref(), input.1),
+            expected.as_ref()
+        );
+    }
 }