+++ src/io/report.rs
@@ -1,52 +1,30 @@
-use color_eyre::{eyre::Context, Report};
+use color_eyre::{eyre::Context, Help, Report};
+use core::convert::TryFrom;
 use csv::Reader;
-use daggy::NodeIndex;
 use std::fs::File;
 
-use crate::{
-    kraken::KrakenReportRecord, kraken::Organism, parser::parse_ident_organism_name,
-    tree::add_root_to_tree, tree::find_correct_parent, tree::IndentOrganism, tree::SpideogTree,
-};
+use crate::{kraken::KrakenReportRecord, tree::IndentOrganism, tree::TaxonomyTree};
 
-pub fn read_kraken_report_tree(
-    reader: &mut Reader<File>,
-) -> Result<(SpideogTree, NodeIndex), Report> {
-    let mut tree = SpideogTree::new();
+pub trait ParseKrakenReport: Sized {
+    fn parse(reader: &mut Reader<File>) -> Result<Self, Report>;
+}
 
-    let mut root: Option<NodeIndex> = None;
-    let mut last_node_id: Option<NodeIndex> = None;
-    let mut new_node_id: NodeIndex;
+impl ParseKrakenReport for TaxonomyTree {
+    fn parse(reader: &mut Reader<File>) -> Result<Self, Report> {
+        let first_record: KrakenReportRecord = reader.deserialize().next().unwrap()?;
+        let origin = IndentOrganism::try_from(first_record)?;
+        let mut taxonomy_tree = TaxonomyTree::new(origin);
 
-    for result in reader.deserialize() {
-        let record: KrakenReportRecord = result.wrap_err("Failed to parse line")?;
+        for result in reader.deserialize() {
+            let record: KrakenReportRecord = result
+                .wrap_err("failed to parse line")
+                .suggestion("make sure that the file is a Kraken2 report")?;
 
-        let (_, (indent, name)) = parse_ident_organism_name(&record.5.as_bytes()).unwrap();
-
-        let organism = Organism {
-            taxonomy_level: record.3,
-            name: String::from_utf8_lossy(name).trim().to_string(),
-            taxonomy_id: record.4,
-        };
-
-        let node = IndentOrganism { indent, organism };
-
-        if root.is_none() {
-            new_node_id = add_root_to_tree(node, &mut tree)?;
-            root = Some(new_node_id);
-            last_node_id = Some(new_node_id);
-        } else if let Some(id) = last_node_id {
-            let parent = find_correct_parent(&node, id, root.unwrap(), &tree)?;
-
-            let (_, new_node_id) = tree.add_child(parent, 1, node);
-            last_node_id = Some(new_node_id);
-        } else {
-            panic!("Kraken tree didn't initialize properly");
+            let node = IndentOrganism::try_from(record)?;
+            let parent = taxonomy_tree.find_valid_parent_for(&node);
+            taxonomy_tree.child(parent, node);
         }
-    }
 
-    if let Some(root) = root {
-        Ok((tree, root))
-    } else {
-        Err(eyre!("Failed to add nodes"))
+        Ok(taxonomy_tree)
     }
 }