+++ src/main.rs
@@ -1,3 +1,7 @@
+// #![warn(clippy::all, clippy::pedantic, clippy::nursery, clippy::cargo)]
+// #![allow(clippy::missing_const_for_fn)]
+// #![allow(clippy::multiple_crate_versions)]
+
 #[macro_use]
 extern crate serde;
 #[macro_use]
@@ -6,8 +10,6 @@ extern crate clap;
 extern crate custom_derive;
 #[macro_use]
 extern crate enum_derive;
-#[macro_use]
-extern crate eyre;
 
 mod bracken;
 mod cli;
@@ -20,77 +22,82 @@ mod utils;
 
 use crate::clap::Clap;
 use cli::{subcommands::Command, Opts};
-use color_eyre::{eyre::Report, eyre::WrapErr};
 
+use color_eyre::eyre::Report;
 use displaydoc::Display;
 use io::{get_output_file_name, read_report_tree, write_tree};
-use kraken::{KrakenReportRecord, Organism};
-use log::info;
+use kraken::KrakenReportRecord;
 use parser::parse_ident_organism_name;
-use std::convert::TryFrom;
 use thiserror::Error;
-use tree::IndentOrganism;
+use tracing::{info, instrument};
 
 #[derive(Display, Error, Debug)]
 #[non_exhaustive]
 pub enum ErrorKind {
-    /// Expected root with no indentation, found indentation level: `{0}`
+    /// expected root with no indentation, found indentation level: `{0}`
     NonZeroIndentRoot(usize),
-    /// No suitable parent found for node `{0}` of indent `{1}` and rank `{2}`
+    /// no suitable parent found for node `{0}` of indent `{1}` and rank `{2}`
     NoSuitableParent(String, usize, taxonomy::TaxonomyRank),
-    /// No node added to the tree
+    /// no node added to the tree
     NoNodeAdded,
-    /// Failed to parse line `{0}`
+    /// failed to parse line `{0}`
     LineParsingError(usize),
-    ///Io error
-    #[error(transparent)]
-    Io(#[from] std::io::Error),
-    ///CSV parser error
-    #[error(transparent)]
-    Csv(#[from] csv::Error),
+    /// failed to parse taxonomy rank offset from `{0}`: `{1}` is not a number (0..9)
+    TaxRankParsingOfffsetNotANumber(String, char), // TODO: find a better solution to this mess
+    /// failed to parse taxonomy rank from `{0}`: found length `{1}` expected 1 or 2
+    TaxRankParsingInvalidLength(String, usize),
+    /// failed to parse taxonomy rank from `{0}`: invalid rank code `{1}` expected R, D, K, P, C, O, F, G, S, U, or -
+    TaxRankParsingInvalidRankCode(String, char),
+    /// failed to parse taxonomy rank from `{0}`: cannot infer previous taxonomy rank from previous records
+    TaxRankParsingCannotInferRank(String),
+    /// node not found
+    NodeNotFound,
+    // ///IO error
+    // #[error(transparent)]
+    // Io(#[from] std::io::Error),
+    // ///CSV parser error
+    // #[error(transparent)]
+    // Csv(#[from] csv::Error)
 }
 
-impl TryFrom<KrakenReportRecord> for IndentOrganism {
-    type Error = Report;
+// Boilerplate: https://github.com/yaahc/color-eyre/blob/master/examples/usage.rs
+// TODO: adjust for use
+// TODO: move to logging.rs?
+fn install_tracing() {
+    use tracing_error::ErrorLayer;
+    use tracing_subscriber::prelude::*;
+    use tracing_subscriber::{fmt, EnvFilter};
 
-    fn try_from(value: KrakenReportRecord) -> Result<Self, Self::Error> {
-        let (_, (indent, name)) = parse_ident_organism_name(value.5.as_bytes()).unwrap();
+    let fmt_layer = fmt::layer().with_target(false);
+    let filter_layer = EnvFilter::try_from_default_env()
+        .or_else(|_| EnvFilter::try_new("info"))
+        .unwrap();
 
-        let organism_tree = Organism {
-            taxonomy_level: value.3,
-            name: String::from_utf8_lossy(name).trim().to_string(),
-            taxonomy_id: value.4,
-        };
-
-        let node = IndentOrganism {
-            indent,
-            organism: organism_tree,
-        };
-
-        Ok(node)
-    }
+    tracing_subscriber::registry()
+        .with(filter_layer)
+        .with(fmt_layer)
+        .with(ErrorLayer::default())
+        .init();
 }
+
+#[instrument]
 fn main() -> Result<(), Report> {
+    install_tracing();
     cli::error::setup_error_hook()?;
 
     let opts: Opts = Opts::parse();
-    opts.logging.setup().wrap_err("Failed to setup logging.")?;
+    // opts.logging.setup().wrap_err("Failed to setup logging.")?;
 
     match opts.command {
         Command::Tree(sub_opts) => {
+            info!("subcommand `tree`");
             // TODO check all files exists, gather errors with eyre
             // https://github.com/yaahc/color-eyre/blob/master/examples/multiple_errors.rs
-            for report in sub_opts.files.reports.iter() {
-                let (tree, root) = read_report_tree(report, sub_opts.headers)?;
+            for report in &sub_opts.files.reports {
+                let tree = read_report_tree(report, sub_opts.headers)?;
                 let output_path = get_output_file_name(report, &sub_opts.prefix);
                 info!("will write output to `{}`", &output_path.display());
-                write_tree(
-                    tree,
-                    root,
-                    &output_path,
-                    &sub_opts.format,
-                    sub_opts.overwrite,
-                )?;
+                write_tree(tree, &output_path, &sub_opts.format, sub_opts.overwrite)?;
             }
         }
     }