+++ src/tree.rs
@@ -1,9 +1,13 @@
+use crate::parse_ident_organism_name;
+use crate::KrakenReportRecord;
+use color_eyre::Report;
+use core::convert::TryFrom;
 use std::fmt::Display;
 
 use daggy::{Dag, NodeIndex, Walker};
-use log::debug;
+use tracing::instrument;
 
-use crate::{kraken::Organism, taxonomy::TaxonomyRank, ErrorKind};
+use crate::kraken::Organism;
 
 #[derive(Debug)]
 pub struct IndentOrganism {
@@ -17,68 +21,76 @@ impl Display for IndentOrganism {
     }
 }
 
-pub fn ancestors(node: NodeIndex, dag: &Dag<IndentOrganism, u32, u32>) -> Vec<NodeIndex> {
-    let mut ancestors_vec = Vec::new();
-    let mut parent_recursion = dag.recursive_walk(node, |g, n| g.parents(n).iter(g).last());
-    while let Some((_, id)) = parent_recursion.walk_next(&dag) {
-        ancestors_vec.push(id);
+impl IndentOrganism {
+    pub fn inferior_indent(&self, than: &Self) -> bool {
+        self.indent < than.indent
     }
-    ancestors_vec
 }
 
-pub fn find_correct_parent(
-    node: &IndentOrganism,
-    last_node_id: NodeIndex,
-    root: NodeIndex,
-    dag: &Dag<IndentOrganism, u32, u32>,
-) -> Result<NodeIndex, ErrorKind> {
-    let last_node = dag.node_weight(last_node_id).unwrap();
+impl TryFrom<KrakenReportRecord> for IndentOrganism {
+    type Error = Report;
 
-    if last_node.indent < node.indent {
-        debug!(
-            "Parent of `{}` is previously added node `{}`",
-            node.organism, last_node.organism
-        );
-        Ok(last_node_id)
-    } else {
-        debug!(
-            "Parent of `{}` is not the previously added node `{}`, searching for a suitable parent",
-            node.organism, last_node.organism
-        );
+    #[instrument]
+    fn try_from(value: KrakenReportRecord) -> Result<Self, Self::Error> {
+        let (_, (indent, name)) = parse_ident_organism_name(value.5.as_bytes()).unwrap();
 
-        let parents = ancestors(last_node_id, dag);
-        let suitable_parent = parents.iter().find(|id| dag[**id].indent < node.indent);
+        let organism_tree = Organism {
+            taxonomy_level: value.3,
+            name: String::from_utf8_lossy(name).trim().to_string(),
+            taxonomy_id: value.4,
+        };
 
-        if let Some(parent_id) = suitable_parent {
-            let parent = dag.node_weight(*parent_id).unwrap();
-            debug!(
-                "Found suitable parent for `{}` => `{}`",
-                node.organism, parent.organism
-            );
-            Ok(*parent_id)
-        } else if node.organism.taxonomy_level <= TaxonomyRank::Domain(9) {
-            Ok(root)
-        } else {
-            Err(ErrorKind::NoSuitableParent(
-                node.organism.name.clone(),
-                node.indent,
-                node.organism.taxonomy_level,
-            ))
+        let node = Self {
+            indent,
+            organism: organism_tree,
+        };
+
+        Ok(node)
+    }
+}
+
+#[derive(Debug, Default)]
+pub struct TaxonomyTree {
+    pub tree: Dag<IndentOrganism, u32, u32>,
+    pub origin: NodeIndex,
+    pub last_node_added_id: NodeIndex,
+}
+
+impl TaxonomyTree {
+    pub fn new(origin: IndentOrganism) -> Self {
+        let mut tree: Dag<IndentOrganism, u32, u32> = Dag::new();
+        let origin = tree.add_node(origin);
+        let last_node_added_id = origin;
+
+        Self {
+            tree,
+            origin,
+            last_node_added_id,
         }
     }
-}
 
-pub fn add_root_to_tree(
-    node: IndentOrganism,
-    dag: &mut Dag<IndentOrganism, u32, u32>,
-) -> Result<NodeIndex, ErrorKind> {
-    if node.indent != 0 {
-        return Err(ErrorKind::NonZeroIndentRoot(node.indent));
+    pub fn child(&mut self, parent: NodeIndex, node: IndentOrganism) -> &mut Self {
+        let edge = 1;
+        let (_, new_node_id) = self.tree.add_child(parent, edge, node);
+        self.last_node_added_id = new_node_id;
+        self
     }
 
-    let root_node_id = dag.add_node(node);
-    Ok(root_node_id)
-}
+    // find a parent with a lower indent value or default to the origin
+    pub fn find_valid_parent_for(&self, organism: &IndentOrganism) -> NodeIndex {
+        let mut parent_id = self.origin; // default value
+        let mut parent_recursion = self
+            .tree
+            .recursive_walk(self.last_node_added_id, |g, n| g.parents(n).iter(g).last());
 
-// TODO: make as struct and impl functions above for it, also store root with it, and other  info (n species, lower taxonomy rank, etc.)
-pub type SpideogTree = Dag<IndentOrganism, u32, u32>;
+        while let Some((_, id)) = parent_recursion.walk_next(&self.tree) {
+            let node = self.tree.node_weight(id).unwrap();
+            if node.inferior_indent(organism) {
+                parent_id = id;
+                break;
+            }
+        }
+
+        parent_id
+    }
+}